Implemented authorization service for role and removed POST for creating users as new user will bew created by register

+++ backend/src/services/user.service.js
@@ -1,14 +1,6 @@
 import * as userRepository from '../repositories/user.repository.js';
 import AppError from '../errors/AppError.js';
-
-export async function createUser(name, age) {
-    const user = {
-        name,
-        age
-    }
-    const savedUser = await userRepository.saveUser(user);
-    return savedUser;
-}
+import * as authorizationService from './authorization.service.js';
 
 export async function getUsers(){
     return await userRepository.getAllUsers();
@@ -22,7 +14,9 @@ export async function getUserById(id){
     return user;
 }
 
-export async function updateUser(id , name , age){
+export async function updateUser(authenticatedUser, id , name , age){
+    authorizationService.ensureOwnerOrAdmin(authenticatedUser, id);
+
     const user = await userRepository.updateUser(id, name, age);
     if(!user) {
         throw new AppError("User Not found", 404);
@@ -31,7 +25,9 @@ export async function updateUser(id , name , age){
 }
 
 
-export async function patchUser(id , updates){
+export async function patchUser(authenticatedUser, id , updates){
+    authorizationService.ensureOwnerOrAdmin(authenticatedUser, id);
+
     const user = await userRepository.patchUser(id, updates);
     if(!user) {
         throw new AppError("User Not found", 404);
@@ -39,10 +35,19 @@ export async function patchUser(id , updates){
     return user;
 }
 
-export async function deleteUser(id){ 
+export async function deleteUser(authenticatedUser, id){ 
+    authorizationService.ensureOwnerOrAdmin(authenticatedUser, id);
+
     const user = await userRepository.deleteUser(id);
     if(!user) {
         throw new AppError("User Not found", 404);
     }
     return user;
+}
+
+export async function getAuthenticatedUser(id) {
+    const authenticatedUser = await userRepository.getUserById(id);
+    if(!authenticatedUser)
+        throw new AppError("Authentication failed", 401);
+    return authenticatedUser;
 }
